requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
pandas==2.2.3
Pillow==11.0.0
Pygments==2.18.0
Requests==2.32.3
scikit_learn==1.4.2
sentence_transformers==3.3.1
Sphinx==8.1.3
sphinx_rtd_theme==3.0.2
torch==2.5.1
transformers==4.46.3
